+++ js/main.js
@@ -1,29 +1,3 @@
-function generateRandomNumber(min, max) {
-  if (min > max) {
-    throw new Error('Min value cant be more max value');
-  }
-  return Math.round(Math.random() * (max - min) + min);
-}
-
-function isLengthCorrect(str, maxLength) {
-  return str.length <= maxLength;
-}
-
-function generateData() {
-  const photos = new Array();
-
-  for (let i = 1; i <= 25; i++) {
-    photos.push({
-      id: i,
-      url: `photos/${i}.jpg`,
-      description: `Photo number ${i}`,
-      likes: generateRandomNumber(15, 200),
-      comments: generateRandomNumber(0, 200),
-    });
-  }
-
-  return photos;
-}
+import { generateData } from "./data";
 
 generateData();
-isLengthCorrect('aaa', 5);
